requirements.txt
# Web Framework
Flask==2.0.1
gunicorn==20.1.0

# Trading APIs
alpaca-trade-api==3.0.0
alpaca-py==0.25.0
polygon-api-client==1.12.0

# Data & Analysis
pandas==1.5.3
numpy==1.24.0
yfinance==0.1.63

# Scraping & HTTP
beautifulsoup4==4.11.0
requests==2.26.0
playwright==1.40.0
playwright-stealth==1.0.6
aiohttp==3.8.0

# LLM & AI
openai==1.0.0
anthropic==0.7.0

# Utilities
python-dotenv==1.0.0
pytz==2023.3
tzdata==2023.3
python-dateutil==2.8.2

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0
